2025/12/30 AbsVal算子  BatchNorm算子 完成未测试  ArgMax demo未可用

+++ src/mat.cpp
@@ -990,7 +990,7 @@ bool CudaMat::empty() const
     return alloc_bytes == 0 || w == 0 || h == 0 || c == 0 || d == 0;
 }
 
-size_t CudaMat::total() const
+int CudaMat::total() const
 {
     if (dims <= 0) return 0;
     size_t t = 1;
